Move edge helpers to a separate extension of TestBase
